Move style of top-level <li> to appropriate section
In the CX skin, <li> are used for both the top level icons
and menu items. The style for the former is written in
such a way that it affects the latter.

This patch moves the top-level styling to the right section
so it doesn't apply to the <li> inside the drop down menu.

Bug: T386899
Change-Id: I14c2ae8315475f9c298a09b85c241c83e8c01e33