Select editor based on site (mobile vs. desktop)
In MediaWiki, mobile mode is generally defined by
MobileFrontend activation, which, in WMF production,
also correspond to the .m subdomain.

This patch uses the mobile mode, as defined by MobileFrontend,
to choose which CX editor to show.

To test locally, select the site mode using the links in
the footer of any wikipage before proceeding to the dashboard.
The links are "Mobile view" on desktop and "Desktop" on mobile.

In WMF production, site selection happens mostly automatically
based on device detection.

Bug: T387004
Change-Id: I1feda0d12a615608cc5defcd1cc586938e7a3dde