Switch to topic definitions from the WikimediaMessages extension
In order to allow the migration in T380825 to move forwards, rely on the
`ArticleTopicFiltersRegistry` class provided by the WikimediaMessages
extension to get a list of topics, instead of hardcoding our own.

The existing `ArticleTopicsDefinition` class is kept so that it can
adapt the format used by WikimediaMessages to the one needed here, and
also provide a convenient way to specify the RL callback (which also
needs to do localisation). Messages that are no longer needed have been
removed.

This patch also adds a hard dependency on the WikimediaMessages
extension. In theory, this would not be necessary, and the extension
could for example fall back to an empty list of topics when
WikimediaMessages isn't available (keep in mind that the topic taxonomy
is Wikipedia-specific anyway). However, that is left for a future
change.

Note for reviewers: this patch has not been tested.

Bug: T387159
Change-Id: I3ba5428553876e2637a38148545476fea8df7683