AX: Display view translation feedback survey to the user
The code utilizes the QuickSurvey extension to display the
feedback survey.

The survey name is used to uniquely identify and load the form.
If the form does not exist, the load operation will be a no-op.

This approach of loading the survey is deprecated and maybe removed
in the future. We still used this approach because it meets our
use-case better:

* Loading QuickSurveys on SpecialPages is not possible atm.
See: T381886
* We need the survey to be displayed based on a user action, in our
case when they view the translation, which is not possible via
configuration.
* Once a survey is filled by a user, by default it is not displayed
to them again.

Related config patch: Ia06c47b5abcdef53b7bc84ca716c9bf56d541907

Bug: T381886
Change-Id: If95755afe48007b383496f191fc4e5731de967e8