Link to the selected tab in the desktop dashboard
When landing on the desktop dashboard without a tab specified
in the URL hash, and if there is no draft translations, the
bookmarked articles are loaded twice and displayed duplicated
as illustrated in the task. This problem is very visible when
switching back and forth between the dashboards using the banners.

This patch doesn't fix the root issue but minimizes its impact
by always including the current tab in the URL hash. This is
done for both the banner and help panel links.

Bug: T387036
Change-Id: If83c15c3476a4a95b94a7bb0f1f53cbae9c06dca