Fix published translation fetching after wiki redirection
When the user starts a translation from the published translation list,
and the target language of the translation is different from the
current wiki language, the user is redirected to the target wiki. When
the user goes back to the dashboard, the selected dashboard tab is the
"published" tab, but instead of getting the list of published
translations, a loading indicator is displayed indefinitely.

This happens because of a wrong check inside "fetchAllTranslations"
method. This patch fixes this check and also moves the state variable
holding the "fetched" status for the draft and published translations,
from the translator vuex state, to the corresponding composable.

Bug: T387110
Change-Id: Ibf45cb36860cc1127f7a3939196fad34b8351b56