AX Instrumentation: Tweak when "view automatic_translation" is logged
We were logging the view event if there was atleast one string
present in the "leadSectionTranslation" variable but that string
could be very small (and hence translated very fast) or a hidden
HTML element.

Updated the code to wait for atleast 3 strings to be loaded before
logging the view event.

Bug: T366149
Change-Id: I23ba754f1bb372d8a2eac71e2ef4c276fe60f7ba