CX UI lib: Update utility classes to support spacing up to 64px
Currently, the spacing utility classes generated are based on the
defined cols number -which is 12, thus the available spacing classes
are for example pa-1 up to pa-12. However, there is no actual
relationship between the number of the grid columns and the spacing
classes.

This patch removes the dependency on the number of grid columns
and creates spacing utility classes for the following numbers:
1, 2, 3, 4, 5, 6, 7, 8, 10, 12, 16 which corresponds to:
4px, 8px, 12px, 16px, 20px, 24px, 28px, 32px, 40px and 64px.

Modifying the newly defined $spacing variable we can add any
spacing that is multiple of 4.

Change-Id: Ie40e0057bece1e1defc6d841531e487295907156