AX mint languages: Rename variable and fix issues
Until now, we use the name "mintLanguages" for the variable
that holds a map between target languages and their supported
source languages. This patch renames this variable to
"mintTargetToSourceLanguages" to avoid confusion, adds a new
"mintSourceToTargetLanguages" computed variable and uses it
to fix a couple of issues inside "confirm" and "translation"
step language selectors.

Change-Id: Iea4695578ad77ec3d0022ae0977d1d6d4230e04d