UD: Remove enabled languages and rely solely on supported language codes
"SectionTranslationTargetLanguages" config variable was initially used
in SX, to limit the target language options, given that the tool was
in a very initial experimental phase. That logic never was updated and
even now, we rely on this configuration variable to check whether a
target language is a valid target language or not.

However, this restriction doesn't really make sense anymore. Given
that the Unified Dashboard is the default for most (if not all) wikis,
the available target languages should be all the languages supported
by cxserver, so that the user can be redirected properly to the wiki
of their selection. If that wiki doesn't support the Unified Dashboard,
the user will just be navigated to the old CX dashboard. This is
something that already happens, and is taken care by the
SpecialContentTranslation.php class.

To simplify our logic, this patch removes the dependency on the
"SectionTranslationTargetLanguages" config parameter in the Unified
Dashboard, and only relies on the language codes supported by
cxserver to provide the available target languages. This will also
allow us to limit the usage of "SectionTranslationTargetLanguages"
to the SX/UD entrypoints only, which makes our work for these
entrypoints easier, as we have to respect fewer restrictions.

Bug: T390076
Change-Id: If12da56a37d866089aa8e389568bc24ba3994198