Frequent languages entrypoint: Compare domain codes instead of languages
Until now, in order to check whether the entrypoint module should be
registered in backend, we compared the codes of the language links from
the OutputPage object, with the language codes defined in the
"SectionTranslationTarget" configuration parameter. Then, inside the JS
module we compared the found missing languages from backend to the
languages defined in "langMap" local storage property.

However, the languages defined in these 3 sources of truth can be
different, especially for the so-called special language codes
(https://meta.wikimedia.org/wiki/Special_language_codes).

This patch changes this logic, and transforms the interlanguage prefixes
(https://en.wikipedia.org/wiki/Special:Interwiki) to domain codes, using
the SiteMapper. Domain codes are consistent regardless of the language
code used for a special language:
e.g. domain code is always "als" for both "gsw" and "als" language
codes, "simple" for both "simple" and "en-simple" language codes and
so on.

To address the issue for good, we also need to update the
"SectionTranslationTargetLanguages" to use the cxserver supported
language codes, that are perfectly supported by SiteMapper.

Finally, the patch removes the current wiki language code from the
available languages, as the article is already present for the
current language and we'll never display a "Missing in X" banner
for it.

Bug: T390300
Change-Id: I0b1b43abf6e8efc4f81d56545ebb02060c55f2cf