CX: Redirect to target wiki if needed, when CX cookie is set
When the user reaches the CX desktop editor a CX cookie containing the
page title and the source/target languages is created. After having a
CX cookie created, if the user tries to visit the CX editor using the
same source title, and language pair, but in a different wiki than the
target wiki, they will get redirected to the mobile editor in the
selected wiki, which leads to a broken workflow.

In the above case, the user should be redirected to the target wiki,
when ContentTranslationTranslateInTarget config parameter is set. This
patch fixes the issue by checking if the current wiki is the same one
defined in the CX cookie target language, and redirects the user to
the target wiki if needed.

Some opportunistic refactoring also done.

Bug: T390934
Change-Id: I683ecb1b9ac27a9ff6403c9ea1569ea714489627
(cherry picked from commit ed3b57cd7476cb460aa85c847508a26bbe41e200)