Produce mobile agnostic URL for redirection
CX has a lot of redirection going on between different tools
and different wikis. The code in SiteMapper was trying to
keep the ".m" if present in an effort to keep the user in the
same mode (desktop or mobile).

I think this cade can be remove and URLs can always be without
the ".m". Mobile mode would still be detected on the server.

Bug: T390923
Change-Id: I760d27975a2aa58d3ca65bb13c50c21b684ce8aa