SXTranslationListLanguageSelector: Fix backdrop color for desktop devices
This patch replaces the "overlayOpacity" prop in MWDialog component with
the "overlayMode" prop, that controls the color mode of the dialog
overlay, accepting either "light" or "dark".

For SXTranslationListLanguageSelector SFC we use the dark overlay mode,
while the light mode is used for all other dialogs in the application.
The only exception is the dialog for the publishing result, which needs
special handling: in this case we override the overlay color in the
component's styles.

Bug: T351734
Change-Id: I2dfc0cbe9fd3ad3fef30616553dedec29408926d