CX: Add "useExistingSectionPublishOption" composable
Currently, when a user selects an existing section to translate, they
are given two options -inside the "Compare contents" step:
- Expand the existing section with the newly translated content.
- Publish the newly translated section as a new, independent section.

In order to properly support this user option from this step during
publishing, we need to store it till the publishing step, and add it
to the publishing payload.

Until now, the implementation in "Compare contents" step was storing
this option in a local variable ("discardedSections") that was not
used anywhere else. Thus, the user option was practically ignored.

This patch adds the "useExistingSectionPublishOption" composable,
that introduces a "existingSectionPublishOption" ref variable that
holds this user option and can also be used inside the publishing
step to access the option.

The patch also fixes the files involved in "Compare contents" step
to support this new logic and removes the old implementation that
uses the "discardSections".

Bug: T388503
Change-Id: I65440d6adc7e9c6eb7bb31d2a779f50e7e9278e8