Update useTranslationSize to support single section time estimate
The useTranslationSize composable provides the information needed to
display the translation time estimate. Currently if a section is
preselected on the confirmer page, the time estimate is shown for
all missing sections, which is a confusing user experience. This
patch ensures that when a section is preselected, the time estimate
shown is only of the single preselected section.

Bug: T387108
Change-Id: I6ad07a8e7fe44c4e3c696568b65d32133c8e8dbe