CX: Make draft save errors more prominent
It can be quite easy to miss the draft save error. I've tried to
make it a little more prominent.

In addition the error that appears after multiple save attempts are
made, is only visible if you scroll up the page.

Change the error message to specify clearly that the save will be
retried again.

Bug: T376531
Change-Id: Ic82a78dc9b72aa3becde2a316b780ad7240c6551