TranslationCorporaStore: Skip insertion if unique key conflicts
Skip inserting corpora record if the unique key constrict conflicts.
This is better, because it allows processing of remaining translation
units.

This approach was also brought up during a previous discussion:
https://phabricator.wikimedia.org/T256229#8207981

Add logging to see how often this occurs and log the content hash to
determine if we are seeing data loss.

Bug: T391311
Change-Id: I5c73cfccdfa083d2000a9453a475cf436a50ea55
Depends-On: Iabc91b665e8de0295f8f13b3557d1392664e6102