CX: Add error handling when converting VE node to HTML
When loading the article:
https://de.wikipedia.org/w/index.php?title=Spezial:Inhalts%C3%BCbersetzung&campaign=interlanguagelink&filter-id=previous-edits&filter-type=automatic&from=en&page=British+Rail+Class+800&to=de
I see the following error in the console when the source article is
being loaded:

   Uncaught TypeError: itemNode is undefined

This appears to be coming from the ve-cite module:
https://gerrit.wikimedia.org/g/mediawiki/extensions/Cite/+/4175e5ea6572237e20e3d169cdfff860370a2ada/modules/ve-cite/ve.dm.MWReferenceNode.js#143

Currently, the user is not informed of this error, and they continue
translating but when trying to save, they get an unrelated error:

  Uncaught TypeError: sectionState is undefined

This happens due to the error mentioned above that causes all
sections don't get loaded. Add error handling so that the rest of
the sections load and user can continue translating the article.

Also add error handling inside when converting a problematic reference
node to DOM elements ("ve.dm.CXReferenceNode.static.toDomElements"), to
skip such elements during template adaptation/translation. Ideally, we
want to properly support such references -even partially-, but still
this is an improvement for now.

Bug: T380195
Change-Id: Ica3b12abe72b02609912fabbccc3392af8311599