RecentSignificantEditStore: minor query optimization
"ORDER BY" causes database to sort the results, as they are not
in the index order. The caller does not seem to care about the order.
Remove that and the limit, given our code maintains 10 entry limit
at insert time already.

Update dozens of comments referring to the limit of 10. This is an
anti-pattern that almost ensures the comments will get out of
date if we ever change the limit or logic. Introduced a const for
the code.

Noticed during audit of T392839.

Change-Id: I1922c9e704f267ed203fa6097c7732d8f39200ba