CX: Wrap publish logic around try...catch
We've received reports about the publish button stuck in a loading
state. See:
https://www.mediawiki.org/wiki/Talk:Content_translation#c-Nattes_%C3%A0_chat-20250503090100-AvatarFR-20250422084000

This will only happen if there is a script error.

Add try..catch block to avoid that from happening and inform user
that an error has occured. Add logging to ensure that we're aware
of such errors.

Bug: T392982
Change-Id: Id2e0121e16573e8a43a77737549ecfcc9b1275b8