CX: Refine status filtering logic when fetching translations from DB
Previously, when we were fetching the draft/published translations to
display in the Unified Dashboard, we used a condition that matched on
either the status of the parent translation (in cx_translations) or
the status of the section translation (in cx_section_translations).

However, that led to weird results: for example when the parent
translation is "draft", we returned all related section translations
as "draft", even if a section translation was deleted.

This patch updates the status filtering condition to better match
the status for parent translations only if there are no related
section translations, or if the joined section translation is a lead
section.

Bug: T393447
Change-Id: I49d6ab0b4ec1bdfd91b50118f7365ab7a55ba8ed