CX: Fix instrumentation for "dashboard_translation_continue" event
There are 3 problematic fields in the event that throw errors when a
translation started on desktop, is continued:

- "translation_id" because we logged the section translation id -which
is missing for desktop-started translations-, instead of the translation
id which should be logged.
- "translation_source_section" and "translation_target_section" fields
because we logged null values instead of skipping them when empty.

This patch fixes all three issues for desktop-started translations.
For translation started on mobile, the event was working properly
already, even for lead section translations, because the section
translation id is present and source and target section titles are
set to "__LEAD_SECTION__" (thus not empty).

Bug: T395415
Change-Id: I35d5b05e5e1bd8b49dfee72de4f6271a137e4334