Introduce usage of DatabaseVirtualDomains
Introduce a virtual domain virtual-cx that will point to the
extensions cluster, and the wikishared database.

With this configuration we should be able to deprecate
ContentTranslationCluster config parameter.

Updated the code in TranslationStore service to use the two new
methods to fetch a connection to the primary or the replica database.

Add a registration hook handler to add a default virtual domain if
it's missing.

I'll replace all other usages of LoadBalancer::getConnection in a
follow up patch.

Bug: T348513
Change-Id: Ia5a6c72e6a1055ec11061d8cbb113ec4ac04dec1