CX: Fix instrumentation for "dashboard_translation_discard" event
There are 3 problematic fields in the event that throw errors when a
translation started on desktop, is continued:

- "translation_id" because we logged the section translation id -which
can be missing-, instead of the translation id which should be logged.
- "translation_source_section" and "translation_target_section" fields
because we logged null values instead of skipping them when empty.

The patch also has some minor fixes like:
- only handling the "delete-translation" even for draft translations,
since published translations cannot be deleted.
- remove unused imports
- fix documentation blocks

Bug: T395415
Change-Id: I36df49a43a5a68ee1b76f0730a92cded0cdd0f10