CX mobile editor: Instrument "editor_segment_edit" event
Also fixed the value of the "isInitialEdit" history state property,
to be based on whether the segment has already been translated.

This fixes the case when the user clicks to edit a template, then
applies the translation without making any edits (by clicking
on the blue button with the "next" arrow inside SX editor). In
this case, the "editor_segment_add" event should be logged, but
nothing is logged until now.

Bug: T396716
Change-Id: Iac7fc82803d503b73b7206c6970d69e9bb24ea80