CX: Add "nextSessionPosition" method to mw.cx.util
Copied all the methods from "translationSessionPosition" file from
the Unified Dashboard codebase to a new file named
"mw.cx.eventlogging.translationSessionPosition.js. That file only
exports the "nextSessionPosition" method. Given that "mw.cx.util"
module is a dependency both for the desktop editor and the unified
dashboard, this method can be used universally to get the next
session position of the instrumented event.

Bug: T395493
Change-Id: I2bf0ed92f330def9b1210d7ac7829deb227b501c