Add link to see more countries in [region]
When a region has more than 7 countries, on the 'all' tab,
show only the first 7 countries and a link to
"See more countries in [region" that takes the user to the
'region' tab with the region in question expanded with all
its countries visible.

This is configured in SXSuggestionFiltersDialog for regions only
but it is implemented as a generic capability in
SXSuggestionFiltersDialogTabGroupContent and SuggestionFilterChip
so it could potentially be used for collections with too many
sub-collections.

Bug: T391230
Change-Id: I090d1649a4dce342d657ebaba8d88b554a4be8e2