CX section publishing: Add support for expanding existing section
This patch adds support for expanding an existing section, to the
"cxpublishsection" API endpoint. More specifically, it adds a new
API parameter named "existingsectiontitle". When this parameter is
present, the published text should be added to the end of the section
that has a title equal to the "existingsectiontitle" param. Also,
the title of this section is updated to the value of the
"targetsectiontitle" parameter.

A new service to encapsulate the section content calculation is
created and registered in ServiceWiring.php.

This functionality will be used by CX, to properly support the
available publishing options (new/expand) for present sections.

Bug: T388503
Change-Id: I7db030a97cf18c319bb74b3f9a72997890ff5517