Use "direct_preselect" event source when starting translation from URL
Currently, when we start a translation from an entrypoint, the event
source used for the "dashboard_translation_start" instrumentation event
is the same used for the "dashboard_open" event, and its based on the
campaign URL parameter that is used by the entrypoint. However, this
is not desired for the "dashboard_translation_start" because:
1. We can identify from which entrypoint the user is coming from the
   "dashboard_open" event
2. The campaign-based event sources are not valid for this event

This will probably also help to identify the root cause of the "null"
event_source that is logged in some cases for the
"dashboard_translation_start" event.

Bug: T395418
Change-Id: I9be6378a7d8925f04823b28f0c98f7f8f02a2bb6