Desktop draft: Check status on section translation if sectiontitle given
Currently, when we are fetching a draft translation to continue, in
desktop editor, the API queries the DB for parent translations with
draft translation status inside "cx_translations" table.

However, when the requested draft translation is a section translation
(i.e. "sourcesectiontitle" API parameter is present), it is the child
section translation inside "cx_section_translations" table, that
should have "draft" translation status, not the parent translation in
"cx_translations" table.

This patch fixes this logic bug that blocks the user from continuing
draft section translations, when they have already published a
translation for that article.

Bug: T394616
Change-Id: I13208770a8fe66be446e744c17b7b0bb4d50996a