CX mobile editor: Do not re-set page/section titles of restored drafts
Currently, the translated title of a draft section is initialized,
every time the "Pick a sentence" step is mounted, even the draft
translation is already restored.

However, the "Pick a sentence" step is mounted, not only when navigating
to it from the dashboard, but also when returning from the
"Edit a segment" step. In such cases, the draft translation is already
initialized, and re-setting the page/section title leads to lost title
edits.

To avoid such issues, this patch modifies the `useDraftTranslationStart`
composable, to only be executed for non-restored drafts.

Bug: T397749
Change-Id: I867c96fa2bdbf406e44c53820748c4af105ef8a7