TranslationCorporaStore: Mark log on insertion as debug
Logging reveals that duplicate entry for key 'cx_corpora_unique'
occurs in different requests, but with same data size and content hash.
See: https://phabricator.wikimedia.org/T391311#10937924

It should be safe to ignore this error.

The log that is added when a translation unit is inserted is quite
frequent. Mark it as "debug" until we investigate T397910.

Bug: T397910
Bug: T391311
Change-Id: I0262447f2acc7f8a4c8bd6349342a3c8fa87c4a5