SXPublishOptionSelector: Use Codex radio buttons instead of MwRadio
Ideally, we should use Codex components for every suitable usecase.
Here we replace MwRadioGroup and MwRadio components from our CX UI
library, with the equivalent CdxField and CdxRadio components.

After the refactoring, the layout is closer to the specifications
of the component, as defined in T252542:
https://phabricator.wikimedia.org/F31819223

Screenshot before refactoring:
https://phabricator.wikimedia.org/F62822438
Screenshot after fix:
https://phabricator.wikimedia.org/F62822441

Change-Id: I1fceb8301e34a06aa5ff4aa66e6414a78bf428fe