CX custom filters: Add "useSuggestionFiltersInstrument" composable
This patch adds a new composable that includes all the methods related
to the instrumentation of the custom suggestion filter events. The
"getSuggestionFilterEventContext" and "getSuggestionFilterEventSource"
methods are moved from their respective files to this composable and
their files are removed.

Tests are added too.

Some opportunistic refactoring also done here.

Change-Id: If4b752b783af90ce3653b527c6ed58af64f889ae