CX: Introduce usePublishTarget composable
Currently, we have 2 separate "state" variables that control the target
of the publishing of a section translation (new section, expand section,
replace section, sandbox section). These variables are "publishTarget"
in Vuex application/state, and "existingSectionPublishOption" in
"useExistingSectionPublishOption".

This patch introduces a new "usePublishTarget" composable that replaces
the usage of the above variables and uniquely determines the target of
the publishing.

Tests are added too.

Bug: T388503
Change-Id: I1efe0f636583e6897a1b9bfb8b445d0a21b6f167