SXPublisher: Properly support "expand" publishing target
Until now, the publishing target options inside the corresponding dialog
(SXPublishOptionSelector SFC) does not support the "expand" option.
At the same time, the message about expected publishing outcome that is
displayed to the user, indicates that the section will be published as
a new section, even when the "expand" publishing target is selected.

This patch addresses the above issues and adds proper support for the
"expand" publishing target option inside the "Preview and Publish" step.

Bug: T260131
Bug: T388503
Change-Id: I1ae5fa0a041fb7c58fd219b00c1df41719961f46