CX: Show error about IP being blocked
VPN is not allowed when editing translations. However, the error info
about this and and other cases such as 'JWT token missing' are not shown
to the user.

This patch surfaces additional info to the user, which might not be
immediately understood by the them, but provides us context whenever the
a report is made against it in Phabricator. This will help us to debug
issues in the future.

Additionally fix the following error that was noticed during debugging:

> Uncaught TypeError: can't access property "remove",
> this.$loadingIndicator is null

Bug: T391552
Bug: T333880
Bug: T368929
Change-Id: Ice7d8ff50a579f5d24c3e124825758e1f3e80630