CX: Clean up 'cxpublishedtranslations' API endpoint
'from' and 'to' parameters are required, so we don't need to check if
they are set. Additionally, "getAllPublishedTranslations" method expects
"$offset" parameter to be integer, and given that this is a "continue"
parameter we can set its type to "integer".

Finally, the endpoint no longer supports empty 'from' and 'to' params,
so the examples that don't include these required parameters are
removed.

Change-Id: Ifaace42852378b15590839caa305e226ac5d8fdd