Refactor "getCXServerTokenPromise" to return native Promise
Until now "getCXServerTokenPromise" returned an AbortablePromise object,
which was also returned by "getCXServerToken" method and sequentially
returned by the MachineTranslationService::translate method. Eventually,
the result of this "translate" method is returned by the
"changeContentSource" method, leading to a TypeError inside
"ve.ui.CXTranslationAction.translate":
```TypeError: promise.catch(...).finally is not a function```

This patch slightly refactors the "getCXServerTokenPromise" method to
return a native Promise, that properly support "finally" method, so
that this issue is fixed.

Bug: T400026
Change-Id: I0506091f501ad6d4f65acce0392022f73839642c