APIQueryCX: Use section translation progress when fetching a section
When translating a section, the front end code calculates the progress
for a specific section and then compares it to the progress of the
entire article translation which is not updated when saving individual
sections.

Hence when a specific section is queried, return the translation
progress saved for that section.

Bug: T376531
Change-Id: Ic25aaa8ebdab2b0d57a3ca8b235c4df52ea8e281