useArticleSearch: Skip search if search input is empty
Currently, when if the search input is empty, we first set the
"search pending" flag to true, and then call the debounced
"searchForArticles" method. That method is asynchronous, and even
if it returns early for empty inputs, it takes some time for the
"search pending" flag to be reset to false (because of its async
nature). This behaviour can cause glitches in the pending indicators.

This patch fixes this issue by avoiding the call to the "search"
method, if the search input is empty.

Bug: T398633
Change-Id: Ie87c03386344d189f5a07e8ff39c088f5de05457