SXSentenceSelector: Add guard to reject navigation if URL params missing
This patch fixes the below case:
1. The user clicks on a mobile entrypoint (e.g. mobile language searcher)
and lands on the confirmation step.
2. The user clicks on the "Start translation" button and is navigated to
the "Pick a sentence" step.
3. The user clicks on the "arrow previous" icon button and is navigated
to the dashboard.
4. The user clicks on the browser back button and is navigated to the
"Pick a sentence" step without a page title, leading to several errors
and a broken experience. One of the issues is the logging of the editor
open event with null translation_source_title.

Adding, the navigation guard prevents to user from being navigated to
the "Pick a sentence" step, to avoid broken behaviour.

Bug: T395417
Change-Id: Ie12d240f0d9972d965f01239468c825b1e4efd52