New section positioning based on section mappings (server-side only)
Determines the new section position for a translated section based on
source-to-target section mappings:
- If the translated section is missing from the target article, find
the next section in source article that is present in both source and
target, and position the target section exactly before that section.
- If the translated section already exists in the target article and
the user selects the "NEW_SECTION" publishing target, insert it
immediately after the existing target section.

Bug: T386131
Change-Id: Ica5529d1d1dfd9ebaa8951815e5db5a33c45b37a