Fetch section titles from API instead of extracting from content
Until now, the section titles for the PageSection models was extracted
from the segmented content fetched from cxserver. This works well for
regular section titles but fails to work for section titles containing
HTML segments. In such cases the section title node HTML can be like
the following example:

<h2 id="170fa6185417191cd41a691e7effcf">
  <span class="cx-segment" data-segmentid="112">
    Empresaria
    <sup
      id="cite_ref-Globo_1-3"
      about="#mwt11"
      class="mw-ref reference"
      data-mw='{"name":"ref","attrs":{"name":"Globo"}}'
      rel="dc:references"
      typeof="mw:Extension/ref"
    >
      <a id="mwVQ" href="./Viviane_Senna#cite_note-Globo-1">
        <span id="mwVg" class="mw-reflink-text">
          <span id="mwVw" class="cite-bracket">[</span>
          1
          <span id="mwWA" class="cite-bracket">]</span>
        </span>
      </a>
    </sup>
  </span>
</h2>

A title like this would be hard and risky to extract from content. For
this reason, this patch adds a "parse" request to fetch the section
titles from the API, and then assign them to the corresponding
PageSection model. This allows us to provide support for section titles
with HTML, too.

Bug: T387427
Change-Id: I0b47ddd579030d643ecb87f50141d73a274f12eb