useCurrentPages: Expose "currentTargetPageTitle" computed variable
The title of the current target page can be available before fetching
the target page metadata. This patch exposes a new computed variable
from the "useCurrentPages" to avoid requiring the current target page
when only its title is needed.

This will help handle some issues in an edge-case workflow (at least),
as described in the linked task.

Bug: T401009
Change-Id: I337c17b516671adc1d09783de52b39291a93c073