Filter non-top-level sections during section title assignment
In a recent patch (I0b47ddd579030d643ecb87f50141d73a274f12eb), we added
a request to "parse" Action API, to fetch the section titles and assign
them to the corresponding sections, that are returned from the page
content segmentation. However, the "parse" API returns the sections of
all levels, while we only need the top-level sections. That leads to
assigning wrong titles to page sections, which breaks the page section
loading, which is based on section title.

This patch fixes this issue by filtering out the non-top-level sections
and using the "number" property of each section, instead of the "index"
property that was previously used. The "number" represents a section's
position among top-level sections, while the "index" counts all sections
(including nested ones).

Bug: T387427
Change-Id: I451b325c9b8be4f6ddce5d0aae7ee2a77b36dffb