Use content size to calculate difficulty level on the frontend
Suggestion difficulty level is current provided by the
recommendation api. That works well when the user selects
a suggestion from the list but in many other cases, like
search and direct preselect, that do not go through
the recommendation api, difficulty information is not
available so some visual cues as well as instrumentation
cannot work properly.

To address this problem, this patch moves the difficulty
level thresholds to the frontend and relies on article
and section sizes provided by the backend. This avoids
thresholds duplication between several systems.

Change-Id: I34b9ee3e23a5e4b3716d9d72a7a1eaec3817438d