CX: Add CXServerClient service for server-side requests to cxserver
This patch adds a new service and a new configuration parameter named
"ContentTranslationCxServerHost" to properly support cxserver requests
from ContentTranslation backend.

The "ContentTranslationCxServerHost" should always be set to a valid
host, otherwise an exception is thrown.

The value of the configuration parameter in production is expected to
be something like: http://localhost:6015.

Bug: T386131
Change-Id: I885cbf3717023c9e23add2118c92dc4002025032