Add appropriate publish options for SX on desktop editor
Until now, when translating a section on desktop, the available
publishing destination options have been the same as when translating
an article. This is misleading for the user and hides from them other
supported options like "EXPAND", etc.

This patch adds support for the different publishing targets and
updates the corresponding dialog to display the appropriate options
every time.

Bug: T392474
Change-Id: Icc39e87b2c448f573e3ace276251a3a80587e843