Add useSectionPresenceStatus composable
Currently, this composable exposes two computed variables:
`isCurrentSectionPresent` and `isPresentLeadSection`. However, only
`isCurrentSectionPresent` is used in the code; `isPresentLeadSection`
will be used in following patches.

`isCurrentSectionPresent` computed variable has been moved from the
`useCurrentSectionSuggestion` composable and is updated to use the
`isPresentLeadSection` computed variable, so that it also accounts for
present lead sections when checking if the current section is present.

Bug: T345845
Change-Id: Ib6b4a7f217f061229d7111ec74186aa27acf2e94