CX desktop editor: Fetch section mappings for all section translations
Until now, section mappings were only fetched for initializing the
publishing target, if it was missing from the URL. Section mappings
fetching was skipped if the publishing target was present in the URL.
However, section mappings are required to properly publishing sections
present in both source and target languages.

This patch fixes this issue by fetching the section mappings for all
section translations, so that we can properly support publishing
present sections on desktop.

Bug: T260131
Change-Id: Ic999ad4c489fa09b087c584aa4e3feef9db502bf