CX: Add "targetsection" URL parameter when redirecting to desktop editor
Currently, when redirecting to the desktop editor for translating a
specific section, we do not set the "targetsection" URL parameter.
Without this URL parameter, the target section title is always set to
the value of source section title, and the user is required to translate
the section title themselves. Even worse, for present sections the user
is not aware which is the original target title of the section they are
translating, and if they translate the source section title to a
different target section title, this will lead to (probably
unintentionally) renaming the original target section after publishing.

To avoid these issues, this patch sets the "targetsection" URL parameter
for all section translations, when redirecting to the desktop CX editor.

Bug: T409270
Bug: T388503
Change-Id: I11f5153c62130cc907e561f12b13fcb1d71160a4